chore(connect): move web module into main package
