+++ packages/connect/package.json
@@ -107,7 +107,9 @@
         "@trezor/trezor-user-env-link": "workspace:*",
         "@types/jws": "^3.2.11",
         "@types/karma": "^6.3.9",
+        "@types/node": "22.13.10",
         "@types/node-fetch": "^2.6.12",
+        "@types/w3c-web-usb": "^1.0.10",
         "babel-loader": "^10.0.0",
         "crypto-browserify": "3.12.0",
         "jest": "29.7.0",
