+++ packages/suite-build/configs/web.webpack.config.ts
@@ -83,11 +83,6 @@ const config: webpack.Configuration = {
                     filename: path.join(baseDir, 'build', route.pattern, 'index.html'),
                 }),
         ),
-        // imports from @trezor/connect in @trezor/suite package need to be replaced by imports from @trezor/connect-web/src/module
-        new webpack.NormalModuleReplacementPlugin(
-            /@trezor\/connect$/,
-            '@trezor/connect-web/src/module',
-        ),
         ...(!isDev ? [new CssMinimizerPlugin()] : []),
     ],
 };
